pyproject.toml
[project]
name = "streamlit-builder"
version = "0.1.0"
description = "AI-powered Streamlit application builder"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.8.0",
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.4",
    "streamlit>=1.42.0",
    "anthropic>=0.45.2",
    "python-dotenv>=1.0.1",
    "tomli-w>=1.2.0",
    "watchfiles>=1.0.4",
    "tomli>=2.2.1",
    "aiohttp>=3.11.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["streamlit_builder/tests"]
python_files = ["test_*.py"]
addopts = "-v"
# Add this to fix the deprecation warning
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests"
]
filterwarnings = [
    "ignore::DeprecationWarning:anthropic.*:",  # Ignore Anthropic deprecation warnings
]

[tool.hatch.build.targets.wheel]
packages = ["streamlit_builder"]
